% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{available_symbols}
\alias{available_symbols}
\title{Available symbols}
\usage{
available_symbols()
}
\value{
A character vector of currently available symbols to be used as the \code{symbol} value in
functions within the package.
}
\description{
Available symbols
}
\examples{
\donttest{
available_symbols()
}
}
