% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitmexr.R
\docType{package}
\name{bitmexr}
\alias{bitmexr}
\title{bitmexr: R Client for the BitMEX Exchange}
\description{
bitmexr provides tools to access the API for the
BitMEX cryptocurrency derivatives exchange https://www.bitmex.com/.
}
\section{See Also}{

\itemize{
\item https://www.bitmex.com/app/apiOverview
\item https://www.bitmex.com/api/explorer/
}
}

