% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_orders.R
\name{delete_all_orders}
\alias{delete_all_orders}
\title{Delete all orders}
\usage{
delete_all_orders(symbol = NULL, filter = NULL, text = NULL, testnet = TRUE)
}
\arguments{
\item{symbol}{string. Optional symbol. If provided, only cancels orders for that symbol.}

\item{filter}{string. Optional filter for cancellation. Use to only cancel some orders,
e.g. '{"side": "Buy"}'.}

\item{text}{string. Optional cancellation annotation. e.g. 'Spread Exceeded'.}

\item{testnet}{logical. Use \code{TRUE} to query the BitMEX testnet platform.
Set to \code{FALSE} by default.}
}
\description{
Delete all orders that have been placed.
}
