% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_orders.R
\name{delete_order}
\alias{delete_order}
\title{Delete order}
\usage{
delete_order(testnet = TRUE, orderID = NULL, clOrdID = NULL, text = NULL)
}
\arguments{
\item{testnet}{logical. Use \code{TRUE} to query the BitMEX testnet platform.
Set to \code{FALSE} by default.}

\item{orderID}{string. Order ID.}

\item{clOrdID}{string. Optional client ID set when placing an order.}

\item{text}{string. Optional cancellation annotation. e.g. 'Spread Exceeded'.}
}
\description{
Delete an order that has been placed.
}
