% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_order.R
\name{edit_order}
\alias{edit_order}
\title{Edit an order}
\usage{
edit_order(
  testnet = TRUE,
  orderID = NULL,
  origClOrdID = NULL,
  clOrdID = NULL,
  orderQty = NULL,
  leavesQty = NULL,
  price = NULL,
  stopPx = NULL,
  pegOffsetValue = NULL,
  text = NULL
)
}
\description{
Edit an order
}
