% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_data.R
\name{individual_trades}
\alias{individual_trades}
\title{Individual trade data}
\usage{
individual_trades(
  symbol = "XBT",
  filter = NULL,
  columns = NULL,
  count = 1000,
  start = NULL,
  reverse = "true",
  startTime = NULL,
  endTime = NULL
)
}
\arguments{
\item{symbol}{Instrument symbol. You can also send a timeframe, e.g. XBT:quarterly}

\item{filter}{Generic table filter. Send JSON key/value pairs, such as {"key": "value"}. You can key on individual fields, and do more advanced querying on timestamps.}

\item{columns}{Array of column names to fetch. If omitted, will return all columns.}

\item{count}{Number of results to fetch.}

\item{start}{Starting point for results.}

\item{reverse}{If true, will sort results newest first.}

\item{startTime}{Starting date filter for results.}

\item{endTime}{Ending date filter for results.}
}
\value{
`individual_trades` returns a tibble containing data for executed trades
}
\description{
Individual trade data
}
