% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_bucket_trades.R
\name{map_bucket_trades}
\alias{map_bucket_trades}
\title{Bucket trade data over an extended period}
\usage{
map_bucket_trades(
  start_date = "2015-09-25 13:00:00",
  end_date = now(tzone = "UTC"),
  binSize = "1d",
  symbol = "XBTUSD",
  partial = "false"
)
}
\arguments{
\item{start_date}{The start date of the sample. Default to the earliest date for api calls}

\item{end_date}{The end date of the sample. Default to today}

\item{binSize}{Time interval to bucket by. Must be one of: 1m, 5m, 1h or 1d.}

\item{symbol}{Instrument symbol. Use \code{available_symbols()} to see valid symbols.}

\item{partial}{If true, will send in-progress (incomplete) bins for the current time period.}
}
\value{
`map_bucket_trades` returns a data.frame containing bucketed trade data for the specified time frame.
 \item{timestamp}{Date and time of trade}
 \item{symbol}{Instrument ticker}
 \item{open}{Opening price for the bucket}
 \item{high}{Highest price in the bucket}
 \item{low}{Lowest price in the bucket}
 \item{close}{Closing price of the bucket}
 \item{trades}{Number of trades executed within the bucket}
 \item{volume}{Volume in USD}
 \item{vwap}{Volume weighted average price}
 \item{lastSize}{Size of the last trade executed}
 \item{turnover}{How many sathoshi were exchanged}
 \item{homeNotional}{BTC value of the bucket}
 \item{foreignNotional}{USD value of the bucket}
}
\description{
\code{map_bucket_trades} uses purrr::map_dfr under the hood to execute multiple API calls.
This is useful when the data you want to return exceeds the maximum 1000 row response limit,
but do not want to have to manually call \code{\link{bucket_trades}} repeatedly.
}
\details{
\code{map_bucket_trades} takes a start and end date, and creates a sequence of start dates
which are passed in to the \code{startTime} parameter in \code{\link{bucket_trades}}.

The length of time between each start time in each API call is determined by the binSize.
For example, "1d" is chosen as the binSize the length of time between start dates will be 1000 days.
If "1h" is chosen, it will be 1000 hours etc.

The function will print the number of API calls being sent and provides a progress bar in the console

Public API requests are limited to 30 per minute. Consequently, the function uses purrr::slowly to ensure this
limit is never reached while sending multiple API requests.
}
\examples{
\dontrun{
# Get hourly bucketed trade data between 2020-01-01 and 2020-02-01

map_bucket_trades(start_date = "2020-01-01", end_date = "2020-02-01", binSize = "1h")
}

}
\references{
\href{https://www.bitmex.com/api/explorer/#!/Trade/Trade_getBucketed}{API Documentation}
}
\concept{bucket_trades}
