% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_trades.R
\name{map_trades}
\alias{map_trades}
\title{Trade data over an extended period}
\usage{
map_trades(
  start_date = "2016-01-01",
  end_date = "2016-01-05",
  symbol = "XBT",
  filter = NULL
)
}
\arguments{
\item{start_date}{The start date of the desired sample}

\item{end_date}{The end date of the desired sample}

\item{symbol}{Instrument symbol. You can also send a timeframe, e.g. XBT:quarterly}

\item{filter}{Generic table filter. Send JSON key/value pairs, such as "{'key':'value'}".}
}
\description{
The map variant of \code{\link{trades}} uses a repeat loop to continually request trade data between two time points.
The function will stop when the `start_date` is greater than `end_date`.
There is a 2 second pause after each API call to ensure the API request limit is not reached when using the function.
}
\details{
Warning! Due to the extremely large number of trades executed on the exchange,
using this function over an extended of time frame will result in an extremely long running process.
For example, during 2019 the exchanged averaged approximately 630000 trades per day,
with a maximum of 2114878 trades executed in a single day.
Obtaining the trade data for this day alone would take over an hour, and users should consider using
\code{\link{map_bucket_trades}} with a small `binSize` value instead.
}
\examples{
\dontrun{

# Get all trades for XBTUSD between 2016-01-01 and 2016-02-01

map_trades(start_date = "2016-01-01", end_date = "2016_02_01", symbol = "XBTUSD")
}

}
\references{
\href{https://www.bitmex.com/api/explorer/#!/Trade/Trade_get}{API Documentation}
}
\concept{trades}
