% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{trades}
\alias{trades}
\title{Trade data}
\usage{
trades(
  symbol = "XBT",
  filter = NULL,
  columns = NULL,
  count = 1000,
  start = NULL,
  reverse = "true",
  startTime = NULL,
  endTime = NULL
)
}
\arguments{
\item{symbol}{Instrument symbol. You can also send a timeframe, e.g. XBT:quarterly}

\item{filter}{Generic table filter. Send JSON key/value pairs, such as "{'key':'value'}".}

\item{columns}{Array of column names to fetch. If omitted, will return all columns.}

\item{count}{Number of results to fetch. Maximum of 1000 (the default) per request.}

\item{start}{Starting point for results.}

\item{reverse}{If true, will sort results newest first (default = "true").}

\item{startTime}{Starting date filter for results.}

\item{endTime}{Ending date filter for results.}
}
\value{
`trades` returns a data.frame containing information for executed trades for the given arguments.
}
\description{
`trades` returns information regarding individual trades that have taken place on the
 exchange for the given symbol / time frame.
}
\details{
By default, `trades` will return the most recent trade data first.
This option can be change by setting `reverse` to "false"
}
\examples{

# Return 1000 most recent trades for symbol "XBTUSD".

trades(symbol = "XBTUSD")

# Use for very specific values.
# Return all trade data executed at 12:15.

trades(symbol = "XBT", filter = "{'timestamp.minute':'12:15'}")

# Also possible to combine more than one filter.

trades(symbol = "XBT", filter = "{'timestamp.minute':'12:15', 'size':10000}")


}
\references{
\href{https://www.bitmex.com/api/explorer/#!/Trade/Trade_get}{API Documentation}
}
\concept{map_trades}
