% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_data.R
\name{bucket_trades}
\alias{bucket_trades}
\title{Bucketed trade data}
\usage{
bucket_trades(
  binSize = "1m",
  partial = NULL,
  symbol = "XBT",
  filter = NULL,
  columns = NULL,
  count = 1000,
  start = NULL,
  reverse = "true",
  startTime = NULL,
  endTime = NULL
)
}
\arguments{
\item{binSize}{Time interval to bucket by. Available options: 1m,5m,1h,1d}

\item{partial}{If true, will send in-progress (incomplete) bins for the current time period.}

\item{symbol}{Instrument symbol. You can also send a timeframe, e.g. XBT:quarterly}

\item{filter}{Generic table filter. Send JSON key/value pairs, such as {"key": "value"}. You can key on individual fields, and do more advanced querying on timestamps.}

\item{columns}{Array of column names to fetch. If omitted, will return all columns.}

\item{count}{Number of results to fetch.}

\item{start}{Starting point for results.}

\item{reverse}{If true, will sort results newest first.}

\item{startTime}{Starting date filter for results.}

\item{endTime}{Ending date filter for results.}
}
\value{
`bucket_trades` returns a tibble containing open high low close data for the specified time interval
}
\description{
Bucketed trade data
}
