% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket_trades.R
\name{bucket_trades}
\alias{bucket_trades}
\title{Bucketed trade data}
\usage{
bucket_trades(
  binSize = "1m",
  partial = "false",
  symbol = "XBT",
  filter = NULL,
  columns = NULL,
  count = 1000,
  start = NULL,
  reverse = "true",
  startTime = NULL,
  endTime = NULL
)
}
\arguments{
\item{binSize}{Time interval to bucket by. Must be one of: 1m, 5m, 1h or 1d.}

\item{partial}{If true, will send in-progress (incomplete) bins for the current time period.}

\item{symbol}{Instrument symbol. You can also send a timeframe, e.g. XBT:quarterly}

\item{filter}{Generic table filter. Send JSON key/value pairs, such as "{'key':'value'}".}

\item{columns}{Array of column names to fetch. If omitted, will return all columns.}

\item{count}{Number of results to fetch. Maximum of 1000 (the default) per request.}

\item{start}{Starting point for results.}

\item{reverse}{If true, will sort results newest first (default = "true").}

\item{startTime}{Starting date filter for results.}

\item{endTime}{Ending date filter for results.}
}
\value{
`bucket_trades` returns a data.frame containing open high low close data for the specified time interval
}
\description{
\code{bucket_trades} retrieves open high low close (OHLC) data for the specified symbol/timeframe.
}
\details{
The API will only return 1000 rows per call. If the desired timeframe requires more than one API call,
consider using \code{\link{map_bucket_trades}}.
}
\examples{

# Return most recent data for symbol "ETHUSD" for 1 hour buckets

bucket_trades(binSize = "1h", symbol = "ETHUSD")

}
\references{
\href{https://www.bitmex.com/api/explorer/#!/Trade/Trade_getBucketed}{API Documentation}
}
\concept{map_bucket_trades}
