% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_orders.R
\name{cancel_order}
\alias{cancel_order}
\title{Cancel order}
\usage{
cancel_order(orderID = NULL, clOrdID = NULL, text = NULL)
}
\arguments{
\item{orderID}{string. Order ID.}

\item{clOrdID}{string. Optional client ID set when placing an order.}

\item{text}{string. Optional cancellation annotation. e.g. 'Spread Exceeded'.}
}
\value{
Returns a \code{data.frame} with details about the order that was cancelled.
See \url{https://www.bitmex.com/api/explorer/#!/Order/Order_cancel} for more information.
}
\description{
Cancel an order that has been placed.
}
\examples{
\dontrun{
# Cancel an order
cancel_order(clOrdID = "myorderid")
}

}
