% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_orders.R
\name{delete_order}
\alias{delete_order}
\title{Delete order}
\usage{
delete_order(testnet = TRUE, orderID = NULL, clOrdID = NULL, text = NULL)
}
\arguments{
\item{testnet}{logical. Use \code{TRUE} to query the BitMEX testnet platform.
Set to \code{FALSE} by default.}

\item{clOrdID}{string. Optional Client Order ID. This clOrdID will come back on the order and any related executions.}

\item{text}{string. Optional order annotation. e.g. 'Take profit'.}
}
\description{
Delete order
}
