% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{get_bitmex}
\alias{get_bitmex}
\title{GET requests}
\usage{
get_bitmex(path, args = NULL, testnet = FALSE, use_auth = FALSE)
}
\arguments{
\item{path}{string. End point for the api.}

\item{args}{A named list containing the}

\item{testnet}{logcial. Use \code{TRUE} access the testnet API, \code{FALSE} to access the live API.}

\item{use_auth}{logcial. Use \code{TRUE} to access private endpoints.}
}
\description{
Use \code{get_bitmex()} to send GET requests. For private endpoints, authentication is required.
}
