% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{get_bitmex}
\alias{get_bitmex}
\title{GET requests}
\usage{
get_bitmex(path, args = NULL, use_auth = FALSE, testnet = FALSE)
}
\arguments{
\item{path}{string. End point for the api.}

\item{args}{A named list containing the}

\item{use_auth}{logcial. Use \code{TRUE} to use apikey with request}

\item{testnet}{logcial. Use \code{TRUE} access testnet.bitmex, \code{FALSE} to access the live site.}
}
\description{
GET requests
}
