% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_trades.R
\name{map_trades}
\alias{map_trades}
\title{Trade data over an extended period}
\usage{
map_trades(
  start_date = "2019-01-01 12:00:00",
  end_date = "2019-01-01 12:15:00",
  symbol = "XBT",
  filter = NULL
)
}
\arguments{
\item{start_date}{The start date of the desired sample}

\item{end_date}{The end date of the desired sample}

\item{symbol}{Instrument symbol. You can also send a timeframe, e.g. XBT:quarterly}

\item{filter}{Generic table filter. Send JSON key/value pairs, such as "\{'key':'value'\}".}
}
\value{
A data.frame containing individual trade information for the specific time period / symbol.
 \item{timestamp}{Date and time of trade}
 \item{symbol}{Instrument ticker}
 \item{side}{Whether the trade was buy or sell}
 \item{size}{Size of the trade}
 \item{price}{Price the trade was executed at}
 \item{tickDirection}{Indicates if the trade price was higher, lower or the same as the previous trade price}
 \item{trdMatchID}{Unique trade ID}
 \item{grossValue}{How many sathoshi were exchanged. 1 satosi = 0.00000001 BTC}
 \item{homeNotional}{BTC value of the trade}
 \item{foreignNotional}{USD value of the trade}
}
\description{
The map variant of \code{\link{trades}} uses a repeat loop to continually request trade data between two time points.
The function will stop when the 'start_date' is greater than 'end_date'.
Given the large number of trades executed per day,
a warning message with a choice to continue, is presented when inputting a date range spanning more than one day.
}
\details{
Warning! Due to the extremely large number of trades executed on the exchange,
using this function over an extended of time frame will result in an extremely long running process.
For example, during 2019 the exchange averaged approximately 630000 trades per day,
with a maximum of 2114878 trades being executed in a single day.
Obtaining the trade data for this day alone would take over an hour, and users should consider using
\code{\link{map_bucket_trades}} with a small 'binSize' (e.g., "1m") instead.
}
\examples{
\dontrun{

# Get all trade data between 2019-05-03 12:00:00 and 2019-05-03 12:15:00

map_trades(start_date = "2019-05-03 12:00:00", end_date = "2019-05-03 12:15:00", symbol = "XBTUSD")

}

}
\references{
\href{https://www.bitmex.com/api/explorer/#!/Trade/Trade_get}{API Documentation}
}
\concept{trades}
