% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{post_bitmex}
\alias{post_bitmex}
\title{POST requests}
\usage{
post_bitmex(path, args = NULL)
}
\arguments{
\item{path}{string. End point for the api.}

\item{args}{A named list containing the}
}
\description{
Use \code{post_bitmex()} to send POST requests. All POST requests require authentication.
}
