% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{trades}
\alias{trades}
\title{Trade data}
\usage{
trades(
  symbol = "XBTUSD",
  filter = NULL,
  columns = NULL,
  count = 1000,
  start = NULL,
  reverse = "true",
  startTime = NULL,
  endTime = NULL
)
}
\arguments{
\item{symbol}{Instrument symbol. Use \code{available_symbols()} to see valid symbols.}

\item{filter}{Generic table filter. Send JSON key/value pairs, such as "\{'key':'value'\}".}

\item{columns}{Array of column names to fetch. If omitted, will return all columns.}

\item{count}{Number of results to fetch. Maximum of 1000 (the default) per request.}

\item{start}{Starting point for results.}

\item{reverse}{If true, will sort results newest first (default = "true").}

\item{startTime}{Starting date filter for results.}

\item{endTime}{Ending date filter for results.}
}
\value{
\code{trades()} returns a data.frame containing information for executed trades for the given arguments.
 \item{timestamp}{Date and time of trade}
 \item{symbol}{Instrument ticker}
 \item{side}{Whether the trade was buy or sell}
 \item{size}{Size of the trade}
 \item{price}{Price the trade was executed at}
 \item{tickDirection}{Indicates if the trade price was higher, lower or the same as the previous trade price}
 \item{trdMatchID}{Unique trade ID}
 \item{grossValue}{How many sathoshi were exchanged. 1 satosi = 0.00000001 BTC}
 \item{homeNotional}{BTC value of the trade}
 \item{foreignNotional}{USD value of the trade}
}
\description{
\code{trades()} retrieves information regarding individual trades that have taken place on the
exchange for the given symbol / time frame.
}
\examples{
\dontrun{
# Return 1000 most recent trades for symbol "XBTUSD".
trades(symbol = "XBTUSD")

# Use filter for very specific values: Return trade data executed at 12:15.
trades(symbol = "XBTUSD",
       filter = "{'timestamp.minute':'12:15'}")

# Also possible to combine more than one filter.
trades(symbol = "XBTUSD",
       filter = "{'timestamp.minute':'12:15', 'size':10000}")
}

}
\references{
\href{https://www.bitmex.com/api/explorer/#!/Trade/Trade_get}{API Documentation}
}
\concept{map_trades}
