% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_dates}
\alias{valid_dates}
\title{Start date of data availability for available symbols}
\usage{
valid_dates(symbol = NULL)
}
\arguments{
\item{symbol}{character string of the instrument
symbol to find start date for.}
}
\value{
A data.frame containing the symbol and date from which data is available
}
\description{
Pass in a symbol from \code{\link[=available_symbols]{available_symbols()}} or no symbol to return dates for all available symbols
}
\examples{
\dontrun{
valid_dates("XBTUSD")
}
}
